refactor: queue drag state handling
