+++ crates/components/src/queue_drag.rs
@@ -1,8 +1,53 @@
-use crate::track_row::{cancel_dragged_queue_track, has_dragged_queue_track};
 use dioxus::document::eval;
 use dioxus::prelude::*;
+use reader::models::Track;
+use std::sync::{Mutex, OnceLock};
 
 pub const RIGHTBAR_DROPZONE_ID: &str = "rightbar-dropzone";
+pub static DRAGGED_QUEUE_TRACK: OnceLock<Mutex<Option<Track>>> = OnceLock::new();
+
+fn dragged_queue_track() -> &'static Mutex<Option<Track>> {
+    DRAGGED_QUEUE_TRACK.get_or_init(|| Mutex::new(None))
+}
+
+pub fn take_dragged_queue_track() -> Option<Track> {
+    dragged_queue_track().lock().ok()?.take()
+}
+
+pub fn has_dragged_queue_track() -> bool {
+    dragged_queue_track()
+        .lock()
+        .map(|guard| guard.is_some())
+        .unwrap_or(false)
+}
+
+pub fn set_dragged_queue_track(track: Track) {
+    if let Ok(mut guard) = dragged_queue_track().lock() {
+        *guard = Some(track);
+    }
+}
+
+pub fn cancel_dragged_queue_track() {
+    clear_dragged_queue_track();
+}
+
+pub fn clear_dragged_queue_track() {
+    if let Ok(mut guard) = dragged_queue_track().lock() {
+        *guard = None;
+    }
+}
+
+pub fn handle_select_click(
+    is_selected: bool,
+    is_selection_mode: bool,
+    on_select: Option<EventHandler<bool>>,
+) {
+    if is_selection_mode {
+        if let Some(handler) = on_select {
+            handler.call(!is_selected);
+        }
+    }
+}
 
 pub fn install_rightbar_drag_handlers() {
     let _ = eval(
